backend/requirements.txt
requests==2.32.4
psycopg2[binary]
Django
djangorestframework
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter==22.1
python-dotenv==0.19.0
gunicorn
djoser
rest_framework_simplejwt
